starting update all project to components view
