+++ package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "scripts": {
     "start": "webpack-dev-server --config ./webpack.config.js",
-    "build": "webpack -p --config ./webpack.config.production.js"
+    "build": "webpack -p --config ./webpack.config.production.js",
+    "analyzer": "./node_modules/webpack/bin/webpack.js --json | ./node_modules/webpack-bundle-size-analyzer/webpack-bundle-size-analyzer"
   },
   "devDependencies": {
     "as-jqplot": "^1.0.8",
@@ -14,11 +15,11 @@
     "babel-loader": "^6.2.4",
     "babel-polyfill": "^6.13.0",
     "babel-preset-es2015": "^6.6.0",
-    "es6-loader": "^0.2.0",
     "bootstrap": "^4.0.0-alpha.3",
     "bootstrap-slider": "*",
     "clean-webpack-plugin": "^0.1.10",
     "css-loader": "^0.18.0",
+    "es6-loader": "^0.2.0",
     "extract-text-webpack-plugin": "^0.8.2",
     "file-loader": "^0.8.4",
     "gulp": "~3.9.1",
@@ -45,6 +46,7 @@
     "tether": "^1.3.3",
     "url-loader": "^0.5.7",
     "webpack": "^1.13.2",
+    "webpack-bundle-size-analyzer": "^2.0.2",
     "webpack-dev-server": "^1.14.1"
   },
   "dependencies": {
