+++ src/components/campaign/photoswipe-ui-default.js
