+++ src/components/capitalRaiseCalculator/model.js
