+++ src/components/capitalRaiseCalculator/styles/style.sass
