+++ src/components/capitalRaiseCalculator/views.js
