+++ src/components/payBackShareCalculator/styles/style.sass
