+++ src/helpers/calculatorHelpers.js
