+++ src/helpers/flyPriceFormatter.js
