+++ src/helpers/shareButtonHelper.js
